+++ package.json
@@ -39,7 +39,7 @@
     "onCommand:type"
   ],
   "qna": "https://vscodevim.herokuapp.com/",
-  "main": "./out/extensionNodeEntry",
+  "main": "./out/extension",
   "contributes": {
     "commands": [
       {
