+++ src/actions/include-all.ts
@@ -1,0 +1,17 @@
+import './base';
+import './operator';
+import './motion';
+import './textobject';
+
+// commands
+import './commands/insert';
+import './commands/actions';
+import './commands/commandLine';
+
+// plugin
+import './plugins/camelCaseMotion';
+import './plugins/easymotion/easymotion.cmd';
+import './plugins/easymotion/registerMoveActions';
+import './plugins/sneak';
+import './plugins/replaceWithRegister';
+import './plugins/surround';
